templates/python/flask/requirements-test.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Caching==2.3.0
marshmallow==3.20.1
python-dotenv==1.0.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.0
redis==5.0.1
celery==5.3.4
requests>=2.32.2
prometheus-flask-exporter==0.23.0
structlog==23.2.0
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
flake8>=6.0.0
bandit>=1.7.0